Refactoring package for MVC
