+++ app/src/main/java/com/example/instagramrecovery/data/GetUserData.java
@@ -1,4 +1,6 @@
-package com.example.instagramrecovery;
+package com.example.instagramrecovery.data;
+
+import com.example.instagramrecovery.presentation.model.RestDataResponse;
 
 import retrofit2.Call;
 import retrofit2.http.GET;
@@ -7,5 +9,6 @@ import retrofit2.http.Query;
 public interface GetUserData {
     @GET("/me/media")
     Call<RestDataResponse> requestGetData(@Query("fields") String field, @Query("access_token") String accessToken);
+
 }
