+++ app/src/main/java/com/example/instagramrecovery/presentation/model/RestDataResponse.java
@@ -1,4 +1,4 @@
-package com.example.instagramrecovery;
+package com.example.instagramrecovery.presentation.model;
 
 import java.util.List;
